pyproject.toml
[project]
name = "quarto-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "patapelada", email = "77882237+patapelada@users.noreply.github.com"},
]
dependencies = ["fastapi[standard]>=0.115.12", "uvicorn>=0.34.3", "python-socketio>=5.13.0", "quarto-lib>=2.1.7"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "dotenv run -- uvicorn quarto_backend.main:app --reload"